+++ js/abstract-view.js
@@ -24,7 +24,7 @@ export default class AbstractView {
     return render(this.template);
   }
 
-  bind(element) {
+  bind() {
 
   }
 }
